wconfig: find config directory
once dotenv.load_dotenv() is run it will overwrite WAVY_CONFIG for the
future. it does therefore not make sense to run it on a conditional or
dependent on what exists in something that is run multiple times.
Anyway, I think it makes sense to let `.env` override WAVY_CONFIG, and
fallback to WAVY_CONFIG if available. This way you get a hybrid where
you can depend on the current directory if you put a `.env` there.

I think this information should go in docs/config.rst with a description
+ spec of the yaml files.